Reorganized src directory and updated file names: - Created the pitch and onset subdirectories to make it easier to keep track of files. - renamed: testOnset.[ch] -> pairTransientDetection.[ch] - renamed: comparison.[ch] -> extractMelodyProcedure.[ch] - tweaked the file permissions of fastGaussTransform.[ch] (for some strange reason these files were previously marked as executable)
