+++ src/extractMelodyProcedure.h
