+++ src/onset/fastGaussTransform.h
