+++ src/onset/onsetStrat.h
@@ -1,4 +1,4 @@
-#include "lists.h"
+#include "../lists.h"
 
 typedef int (*OnsetStrategyFunc)(float** AudioData, int size, int dftBlocksize,
 			int samplerate, intList* onsets);
