+++ src/pitch/BaNaDetection.c
@@ -3,7 +3,7 @@
 #include "findpeaks.h"
 #include "findCandidates.h"
 #include "candidateSelection.h"
-#include "lists.h"
+#include "../lists.h"
 #include "BaNaDetection.h"
