+++ src/pitch/findCandidates.h
@@ -1,4 +1,4 @@
-#include "lists.h"
+#include "../lists.h"
 struct orderedList calcCandidates(float* peaks, int numPeaks);
 float calcM(float f_i, float f_j);
 distinctList* distinctCandidates(struct orderedList* candidates,
