+++ src/pitch/pitchStrat.c
