+++ src/pitch/pitchStrat.h
